requirementss.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
async-timeout==4.0.3
bidict==0.22.1
blinker==1.6.2
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
dnspython==2.6.1
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-DebugToolbar==0.14.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-redis==0.4.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
flatbuffers==23.5.26
fsspec==2024.2.0
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
geventhttpclient==2.0.11
greenlet==2.0.2
grpcio==1.62.0
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
libclang==16.0.6
locust==2.24.0
Mako==1.3.2
Markdown==3.5.2
MarkupSafe==2.1.3
numpy==1.25.2
pycountry==22.3.5
python-dotenv==1.0.1
python-engineio==4.6.1
python-socketio==5.8.0
python-version==0.0.2
PyYAML==6.0.1
pyzmq==25.1.2
redis==5.0.3
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
stripe==6.2.0
typing_extensions==4.10.0
urllib3==2.0.4
Werkzeug==2.3.7
wrapt==1.14.1
WTForms==3.1.2
